requirements.txt
joblib==1.0.1
numpy==1.20.3
opencv-python==4.5.1.48
Pillow==8.2.0
ruamel.yaml==0.17.4
ruamel.yaml.clib==0.2.2
scikit-learn==0.24.2
scipy==1.6.3
threadpoolctl==2.1.0
torch==1.8.0
torchvision==0.9.1
typing-extensions==3.10.0.0
